Moved Stormstrike constants to a constants.tsx
