+++ src/parser/shaman/enhancement/constants.tsx
@@ -1,0 +1,13 @@
+import SPELLS from 'common/SPELLS/shaman';
+
+export const STORMSTRIKE_CAST_SPELLS = [
+  SPELLS.STORMSTRIKE_CAST,
+  SPELLS.WINDSTRIKE_CAST,
+];
+
+export const STORMSTRIKE_DAMAGE_SPELLS = [
+  SPELLS.STORMSTRIKE_ATTACK,
+  SPELLS.STORMSTRIKE_ATTACK_OFFHAND,
+  SPELLS.WINDSTRIKE_ATTACK,
+  SPELLS.WINDSTRIKE_ATTACK_OFFHAND,
+];
